requirements_test.txt
ruff==0.9.7
mypy==1.15.0
pytest==8.3.4
pytest-asyncio==1.3.0
pytest-homeassistant-custom-component==0.13.316
pytest-cov==6.0.0
voluptuous==0.15.2
aiohttp==3.11.12
